[16.0][ADD] disbursement_accounting_kmitl, disbursement_finance_kmitl: DR bill and payment bridges
Phase 4 and Phase 5 of the disbursement app split plan. Move all bill
and payment creation from the disbursement core to two dedicated
bridges so DR core only owns the approval lifecycle.

disbursement_accounting_kmitl (depends on disbursement, accounting_kmitl):
- pipeline_status selection_add: bill_draft, bill_posted
- _compute_pipeline_status override: read bill state to advance the
  status once the DR is approved
- Fields moved from core: bill_ids, bill_count, bill_draft_count,
  bill_status_display
- Methods moved from core: _compute_bill_count, _create_bill,
  _create_single_bill, _create_multi_bills, _prepare_bill_vals,
  _prepare_bill_line_vals, _apply_wht_to_bill, action_view_bill,
  action_create_bill
- account.move.disbursement_request_id moves here together with the
  whole models/account_move.py file
- View: adds the Create Bill button (gated by
  group_accounting_kmitl_user) and a Bills smart button to the DR form
- action_cancel override: block cancel when any bill is posted and
  cancel draft bills before chaining to super()

disbursement_finance_kmitl (depends on disbursement, finance_kmitl,
disbursement_accounting_kmitl):
- pipeline_status selection_add: payment_draft, payment_posted, done
- _compute_pipeline_status override: read bill payment_state and
  related payments to advance the status further
- Fields moved from core: payment_ids, payment_count,
  payment_status_display, payment_move_ids, payment_move_count
- Methods moved from core: _compute_payment_ids, action_create_payment
  (with WHT calc + deferred reconciliation), action_view_payments,
  action_view_payment_moves
- models/account_payment.py placeholder for future hooks
- View: adds the Create Payment button (gated by
  group_finance_kmitl_user) and Payments / Payment Moves smart buttons
- action_cancel override: block cancel when bills already have
  payments, then chain to super()

disbursement core:
- Drops the bill and payment fields, computes, and action methods that
  moved to the bridges along with both account_move.py and
  account_payment.py files (and their imports).
- action_cancel now handles only the core concerns (cancel guard +
  budget commitment cleanup). Bridges layer on their checks via super().
- Removes the Create Bill / Create Payment buttons and the smart
  buttons from the DR form so they live in the bridges instead.
