+++ disbursement/models/__init__.py
@@ -1,7 +1,5 @@
 # License LGPL-3.0 or later (https://www.gnu.org/licenses/lgpl).
 
-from . import account_move
-from . import account_payment
 from . import disbursement_request
 from . import disbursement_request_line
 from . import budget_commitment
