+++ disbursement/views/disbursement_request_views.xml
@@ -34,13 +34,6 @@
                     <button name="action_sign" string="Sign" type="object" class="oe_highlight" attrs="{'invisible': [('state', '!=', 'submitted')]}" groups="disbursement.group_disbursement_head" />
                     <button name="action_validate" string="Validate" type="object" class="oe_highlight" attrs="{'invisible': [('state', '!=', 'signed')]}" confirm="Are you sure you want to validate this request?" groups="disbursement.group_disbursement_officer" />
                     <button name="action_approve" string="Approve" type="object" class="oe_highlight" attrs="{'invisible': [('state', '!=', 'verified')]}" confirm="Are you sure you want to approve? This will obligate and consume the budget." groups="disbursement.group_disbursement_manager" />
-                    <button name="action_create_bill" string="Create Bill" type="object" class="oe_highlight" attrs="{'invisible': ['|', ('state', '!=', 'approved'), ('bill_count', '>', 0)]}" groups="disbursement.group_disbursement_officer" />
-                    <button name="action_create_payment"
-                            string="Create Payment"
-                            type="object"
-                            class="oe_highlight"
-                            attrs="{'invisible': ['|', ('pipeline_status', '!=', 'bill_posted'), ('payment_count', '>', 0)]}"
-                            groups="account.group_account_invoice"/>
                     <button name="action_draft" string="Reset to draft" type="object" class="btn btn-secondary" attrs="{'invisible': [('state', 'not in', ('submitted', 'signed', 'cancel'))]}" groups="disbursement.group_disbursement_manager" />
                     <button name="action_cancel" string="Cancel" type="object" attrs="{'invisible': ['|', ('state', 'in', ('draft', 'cancel')), ('pipeline_status', 'in', ('bill_posted', 'payment_draft', 'payment_posted', 'done'))]}" groups="disbursement.group_disbursement_officer" />
                     <field name="state" widget="statusbar" statusbar_visible="draft,submitted,signed,verified,approved" />
@@ -63,32 +56,6 @@
                 </div>
                 <sheet>
                     <div class="oe_button_box" name="button_box">
-                        <button name="action_view_bill" type="object" class="oe_stat_button" icon="fa-pencil-square-o" attrs="{'invisible': [('bill_count', '=', 0)]}">
-                            <div class="o_field_widget o_stat_info">
-                                <span class="o_stat_text">ใบตั้งหนี้ <field name="bill_count" nolabel="1" /></span>
-                                <span class="o_stat_text text-muted small"><field name="bill_status_display" nolabel="1" /></span>
-                            </div>
-                        </button>
-                        <field name="bill_draft_count" invisible="1" />
-                        <button name="action_view_payments"
-                                type="object"
-                                class="oe_stat_button"
-                                icon="fa-money"
-                                attrs="{'invisible': [('payment_count', '=', 0)]}">
-                            <div class="o_field_widget o_stat_info">
-                                <span class="o_stat_text">รายการจ่ายเงิน <field name="payment_count" nolabel="1" /></span>
-                                <span class="o_stat_text text-muted small"><field name="payment_status_display" nolabel="1" /></span>
-                            </div>
-                        </button>
-                        <button name="action_view_payment_moves"
-                                type="object"
-                                class="oe_stat_button"
-                                icon="fa-book"
-                                attrs="{'invisible': [('payment_move_count', '=', 0)]}">
-                            <div class="o_field_widget o_stat_info">
-                                <span class="o_stat_text">รายการล้างหนี้ <field name="payment_move_count" nolabel="1" /></span>
-                            </div>
-                        </button>
                         <button name="open_preview" type="object" class="oe_stat_button" icon="fa-globe">
                             <div class="o_field_widget o_stat_info">
                                 <span class="o_stat_text">Preview</span>
