+++ disbursement_accounting_kmitl/models/account_move.py
@@ -13,4 +13,3 @@ class AccountMove(models.Model):
         index=True,
         copy=False,
     )
-
