+++ disbursement_accounting_kmitl/views/disbursement_request_views.xml
@@ -1,0 +1,37 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+
+    <!-- Form view: add Create Bill button + Bills smart button -->
+    <record id="view_disbursement_request_form_inherit_accounting" model="ir.ui.view">
+        <field name="name">disbursement.request.form.inherit.accounting_kmitl</field>
+        <field name="model">disbursement.request</field>
+        <field name="inherit_id" ref="disbursement.view_disbursement_request_form"/>
+        <field name="arch" type="xml">
+            <!-- Create Bill button placed after Approve -->
+            <xpath expr="//button[@name='action_approve']" position="after">
+                <button name="action_create_bill"
+                        string="Create Bill"
+                        type="object"
+                        class="oe_highlight"
+                        attrs="{'invisible': ['|', ('state', '!=', 'approved'), ('bill_count', '>', 0)]}"
+                        groups="accounting_kmitl.group_accounting_kmitl_user"/>
+            </xpath>
+
+            <!-- Bills smart button -->
+            <xpath expr="//div[@name='button_box']" position="inside">
+                <button name="action_view_bill"
+                        type="object"
+                        class="oe_stat_button"
+                        icon="fa-pencil-square-o"
+                        attrs="{'invisible': [('bill_count', '=', 0)]}">
+                    <div class="o_field_widget o_stat_info">
+                        <span class="o_stat_text">ใบตั้งหนี้ <field name="bill_count" nolabel="1" /></span>
+                        <span class="o_stat_text text-muted small"><field name="bill_status_display" nolabel="1" /></span>
+                    </div>
+                </button>
+                <field name="bill_draft_count" invisible="1" />
+            </xpath>
+        </field>
+    </record>
+
+</odoo>
